minor db error path refactor + methods into modules
